refactor: colocate server actions file
